move exchange module content to exchange package and the interface to a new module
